requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
pillow==10.0.0
tqdm==4.65.0
albumentations==1.3.1
gradio==3.44.0
pyyaml==6.0.1
streamlit==1.24.0